Add files via upload
XYZ Corp. - The Problem and Objective
The text files contain complete loan data for all loans issued by XYZ Corp. through 2007-2015. The data contains the indicator of default, payment information, credit history, etc. The data should be divided into train (June 2007 - May 2015) and out-of-time test (June 2015 - Dec 2015) data. You will have use the training data to build models/analytical solution and finally apply it to test data to measure the performance and robustness of the models. You should use the variable: 'issue_d' to divide the data in the above time periods, the variable is in - format.
Data and Problem Details:
Objective: To build a data model to predict the probability of default, and choose a cut-off based on what you feel is suitable. Alternatively you can also use a modelling technique which gives binary output
Our Plan to Handle this Data:
We will first focus on Importing the data, Understanding it Better do Clean-up of data and later we will split the train and test files based on the criteria given above and then we will build model.